requirements.txt
argcomplete==3.6.2
asgiref==3.8.1
asttokens==2.4.1
attrs==25.1.0
Beaker==1.12.1
beautifulsoup4==4.13.4
blivet==3.12.1
blivet-gui==2.6.0
Brlapi==0.8.6
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
comm==0.2.2
crypt_r==3.13.1
cryptography==44.0.0
cupshelpers==1.0
dasbus==1.7
dbus-python==1.3.2
dbus_next==0.2.3
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
distlib==0.4.0
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.1
django-filter==25.1
django-redis==6.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.3
dnf==4.23.0
drf-yasg==1.21.10
executing==2.1.0
fedora-third-party==0.10
file-magic==0.4.0
filelock==3.18.0
fros==1.1
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
gunicorn==23.0.0
humanize==4.12.0
idna==3.10
inflection==0.5.1
ipykernel==6.29.5
ipython==8.29.0
iso639==0.1.4
jedi==0.19.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
langtable==0.0.69
libcomps==0.1.21
libdnf==0.74.0
libvirt-python==11.0.0
louis==3.33.0
lxml==5.3.2
Mako==1.2.3
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mercurial==6.9.5
nest-asyncio==1.6.0
nftables==0.1
oauthlib==3.3.1
olefile==0.47
packaging==24.2
parso==0.8.4
Paste==3.10.1
perf==0.1
pexpect==4.9.0
pid==2.2.3
pillow==11.1.0
pipx==1.7.1
platformdirs==4.2.2
ply==3.11
productmd==1.45
progressbar2==4.5.0
promise==2.3
prompt_toolkit==3.0.48
psutil==6.1.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pwquality==1.4.5
pycairo==1.25.1
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.4
pyenchant==3.2.2
Pygments==2.18.0
PyGObject==3.50.0
pyinotify==0.9.6
PyJWT==2.9.0
pykickstart==3.62
pyOpenSSL==25.0.0
pyparted==3.13.0
PySocks==1.7.1
python-augeas==1.2.0
python-dateutil==2.8.2
python-decouple==3.8
python-linux-procfs==0.7.3
python-meh==0.52
python-pam==2.0.2
python-utils==3.9.0
python3-openid==3.2.0
pytz==2025.2
pyudev==0.24.3
pyxdg==0.27
PyYAML==6.0.2
pyynl @ file:///builddir/build/BUILD/kernel-6.15.8-build/kernel-6.15.8/linux-6.15.8-200.fc42.x86_64/tools/net/ynl
pyzmq==26.2.0
RapidFuzz==3.11.0
redis==6.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-file==2.0.0
requests-ftp==0.3.1
requests-oauthlib==2.0.0
rpds-py==0.25.0
rpm==4.20.1
rpmautospec==0.8.1
rpmautospec-core==0.1.5
selinux @ file:///builddir/build/BUILD/libselinux-3.8-build/libselinux-3.8/src
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.8-build/selinux-3.8/python/sepolicy
setools==4.5.1
setuptools==74.1.3
simpleaudio==1.0.4
simpleline==1.9.0
six==1.17.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sos==4.9.1
soupsieve==2.7
sqlparse==0.5.3
stack-data==0.6.3
systemd-python==235
text-unidecode==1.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
ufw==0.35
uritemplate==4.2.0
urllib3==2.3.0
userpath==1.9.2
virtualenv==20.32.0
wcwidth==0.2.13
whitenoise==6.9.0
xkbregistry==0.3
drf-yasg==1.21.7
